chg: move test_xlsform_headers.py tests into test_sheet_columns.py
- these test modules are about the same topic
